Move fetcher utils to config.
