+++ config/datadir_test.go
@@ -1,4 +1,4 @@
-package fetcher
+package config
 
 import (
 	"fmt"
